Move tp_tool tests into tests/pootle_tp/utils
